vc-pitchdeck-screening/requirements.txt
pymupdf
openai
tavily-python
sentence-transformers
scikit-learn
python-dotenv
